[*] rework trainer system (all functions are in gan/<gan_name>/trainer)
